pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.63.0"
gunicorn = "20.0.4"
uvicorn = {version = "0.13.4", extras = ["standard"]}
ndjson = "^0.3.1"
aiolinebot = "^0.4.1"
line-bot-sdk = "^2.1.0"
torch = "^1.10.2"
sentencepiece = "^0.1.96"
gensim = "^4.1.2"
Janome = "^0.4.2"
mysql-connector-python = "^8.0.28"
ginza = "^5.1.0"
ja-ginza = "^5.1.0"
spacy = "^3.2.3"
transformers = "^4.17.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.12b0"
isort = "^5.10.1"
pre-commit = "^2.16.0"
flake8 = "^4.0.1"
flake8-bugbear = "^21.11.29"
pyproject-flake8 = "^0.0.1-alpha.2"
fugashi = "^1.1.2"
ipadic = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py39']
line-length = 119

[tool.isort]
profile = "black"
line_length = 119

[tool.flake8]
max-line-length = 119
max-complexity = 10
ignore = "E203,E266,W503,"

[tool.mypy]
python_version = "3.9"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true
show_error_context = true
